+++ NEWS.md
@@ -1,3 +1,13 @@
+# data.io 1.4.0
+
+-   Dependency to {svBase} added. Now the `default_dtx()` function is used to output a data frame object in the user-preferred class (data.frame, data.table or tibble tbl_df) with `read()`.
+
+-   The **dataframe** object class is deprecated. Consequently, `read()` does not output **dataframe** objects any more and `as_dataframe()` or `is_dataframe()` and similar functions are also deprecated. The `as_dataframe=` argument of `read()` is also deprecated now.
+
+-   There is now a list completion for `read$<tab>` and `write$<tab>`.
+
+-   The `read()` function can now download file (and sidecar file) directly for all types (previously, only for functions that accepted it internally, like `readr::read_csv()` but not for `readxl::read_xls()` for instance). The `cache_file=` argument allows to define a file to cache this download. If the file exists, the data are not redownloaded again. The `read()`function also tries to download a possible sidecar file from the URL + ".R" if `sidecar_file = TRUE`.
+
 # data.io 1.3.1
 
 -   Example using `.csv.tar` and `.csv.tar.gz` are eliminated because more recent `readr::read_csv()` function does not seem to handle Tar archives anymore (observed with {readr} version 2.1.2).